requirements.txt
#Core data processing
xarray>=2023.1.0
netCDF4>=1.6.0
pandas>=1.5.0
numpy>=1.24.0
geopandas>=0.12.0

#Visualization
plotly>=5.0.0

#Web Dashboard
streamlit>=1.28.0
streamlit-audiorec>=0.1.3

#ML/Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
chromadb>=0.5.0

#API Framework
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0

#Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9

#--- LLM Integration (Updated for OpenAI) ---
langchain>=0.1.20
langchain-openai>=0.1.8
langchain-community>=0.0.38
langchain-core>=0.1.52
openai>=1.0.0

#Web scraping
beautifulsoup4>=4.11.0

#Development tools
python-dotenv>=1.0.0
pyngrok>=6.0.0

#Time handling
cftime>=1.6.0

#HTTP requests (for API calls)
requests>=2.28.0

# --- Optional: MCP (Model Context Protocol) ---
# These are optional; the app falls back to in-process tool calls if MCP SDK/server not present
mcp>=0.1.0 ; python_version>="3.9"
websockets>=11.0.0

# Optional: fastmcp runtime for running the MCP server
fastmcp>=2.12.0